[wg] move wireguard tunnel methods into the wireguard.Network abstraction
